backend/requirements.txt
flask==3.0.0
werkzeug==3.0.1
pandas==2.1.4
networkx==3.2.1
openpyxl==3.1.2
rdkit-pypi==2022.9.5
gunicorn==21.2.0
flask-cors==4.0.0
authlib==1.2.1
python-dotenv==1.0.0
numpy==1.26.4
requests==2.31.0